[fix] meal 예외처리 추가

+++ src/main/java/Togefit/server/service/MealService.java
@@ -7,6 +7,8 @@ import Togefit.server.model.meal.*;
 import Togefit.server.repository.Meal.MealArrayRepository;
 import Togefit.server.repository.Meal.MealArticleRepository;
 import Togefit.server.repository.Meal.MealRepository;
+import Togefit.server.response.error.CustomException;
+import Togefit.server.response.error.Error;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,7 +28,22 @@ public class MealService {
         this.mealArticleRepository = mealArticleRepository;
     }
 
+    private void checkMeal(Meal[][] meals){
+        for(int i = 0 ; i < meals.length; i++){
+            for(int j = 0 ; j < meals[i].length; j++){
+                if(meals[i][j].getFoodName().length() == 0){
+                    throw new CustomException(new Error("음식 이름이 반드시 필요합니다."));
+                }
+                if(meals[i][j].getQuantity() < 0){
+                    throw new CustomException(new Error("음식의 양은 반드시 양수여야 합니다."));
+                }
+            }
+        }
+    }
+
     public void saveMeal(Meals mealInfo, String userId){
+        checkMeal(mealInfo.getMeals());
+
         MealArticle article = new MealArticle(userId, Calendar.getInstance());
         mealArticleRepository.save(article);
         Long articleId = article.getId();
@@ -44,8 +61,17 @@ public class MealService {
     }
 
     @Transactional
-    public void deleteMeal(Long articleId){
+    public void deleteMeal(Long articleId, String userId){
         // mealArticle -> meal -> mealArray 순서로 삭제
+        Optional<MealArticle> findArticle = mealArticleRepository.findById(articleId);
+
+        if(findArticle.isEmpty()){
+            throw new CustomException(new Error("해당 아티클을 찾지 못했습니다."));
+        }
+
+        if(!findArticle.get().getUserId().equals(userId)){
+            throw new CustomException(new Error(403, "작성자만 삭제할 수 있습니다."));
+        }
 
         mealArticleRepository.deleteById(articleId);
 
@@ -69,6 +95,10 @@ public class MealService {
     public MealInfoByArticleId getMealArticle(Long articleId){
         MealInfoByArticleId article = new MealInfoByArticleId();
         Optional<MealArticle> findMealArticle = this.findOne(articleId);
+
+        if(findMealArticle.isEmpty()){
+            throw new CustomException(new Error("해당 아티클을 찾지 못했습니다."));
+        }
 
         // 아이디 설정
         article.setUserId(findMealArticle.get().getUserId());
@@ -108,6 +138,10 @@ public class MealService {
     }
 
     public Object[] getPagenation(String userId, int limit, int reqNumber, int year, int month){
+        if(month < 1 || month > 12){
+            throw new CustomException(new Error("월의 범위는 1~12입니다."));
+        }
+
         PageRequest pageRequest = PageRequest.of(reqNumber, limit);
 
         List<MealArticle> articles =
@@ -128,6 +162,16 @@ public class MealService {
     public void saveMealOne(MealOne meal, String userId){
         // mealArray에 저장 -> 생성된 id(mealGroupId)와 articleID로 meal에 저장
         Long mealArticleId = meal.getMealArticleId();
+        Optional<MealArticle> findArticle = findOne(mealArticleId);
+
+        if(findArticle.isEmpty()){
+            throw new CustomException(new Error("해당 아티클을 찾지 못했습니다."));
+        }
+
+        if(!findArticle.get().getUserId().equals(userId)){
+            throw new CustomException(new Error(403, "작성자만 추가할 수 있습니다."));
+        }
+
         MealArray mealArray = new MealArray();
         mealArrayRepository.save(mealArray);
 
@@ -136,7 +180,19 @@ public class MealService {
         saveMeals(meal.getMeals(), mealGroupId, mealArticleId);
     }
 
+    private void checkMealOne(MealInfo[] meal){
+        for(int i = 0 ; i < meal.length; i++){
+            if(meal[i].getFoodName().length() == 0){
+                throw new CustomException(new Error("음식 이름이 반드시 필요합니다."));
+            }
+            if(meal[i].getQuantity() < 0){
+                throw new CustomException(new Error("음식의 양은 반드시 양수여야 합니다."));
+            }
+        }
+    }
+
     public void saveMeals(MealInfo[] meal, Long mealGroupId, Long mealArticleId){
+        checkMealOne(meal);
         for(int i = 0 ; i < meal.length; i++){
             Meal one = new Meal(
                     mealGroupId,
@@ -149,10 +205,14 @@ public class MealService {
     }
 
     @Transactional
-    public void updateMealOne(MealOne meal){
+    public void updateMealOne(MealOne meal, String userId){
         Long mealGroupId = meal.getMealListId();
 
         List<Meal> meals = mealRepository.findByMealGroupId(mealGroupId);
+
+        if(meals.size() == 0){
+            throw new CustomException(new Error("해당 아티클을 찾지 못했습니다."));
+        }
         Long mealArticleId = meals.get(0).getArticleId();
 
         mealRepository.deleteByMealGroupId(mealGroupId);
@@ -161,7 +221,18 @@ public class MealService {
     }
 
     @Transactional
-    public void deleteMealOne(Long id){
+    public void deleteMealOne(Long id, String userId){
+        List<Meal> meals = mealRepository.findByMealGroupId(id);
+
+        if(meals.size() == 0){
+            throw new CustomException(new Error("해당 아티클을 찾지 못했습니다."));
+        }
+
+        Long articleId = meals.get(0).getArticleId();
+        Optional<MealArticle> one = mealArticleRepository.findById(articleId);
+        if(!one.get().getUserId().equals(userId)){
+            throw new CustomException(new Error(403, "작성자만 삭제할 수 있습니다."));
+        }
         mealRepository.deleteByMealGroupId(id);
         mealArrayRepository.deleteById(id);
     }
